+++ src/components/super/i-static-page/modules/ss-helpers/tags.js
@@ -27,15 +27,14 @@ const
 	{needInline} = include('src/components/super/i-static-page/modules/ss-helpers/helpers');
 
 const
-	nonce = csp.nonce(),
-	nonceAttr = {toString: () => nonce, escape: false, interpolate: false};
+	nonce = csp.nonce();
 
 const defAttrs = {
 	nonce: nonce ? [`window['${csp.nonceStore()}']`] : undefined
 };
 
 const defInlineAttrs = {
-	nonce: nonce != null && csp.postProcessor ? nonceAttr : undefined
+	nonce: (nonce != null && csp.postProcessor) ? nonce : undefined
 };
 
 exports.getScriptDecl = getScriptDecl;
@@ -100,7 +99,7 @@ function getScriptDecl(lib, body = '') {
 		const attrsObj = {
 			src: lib.src,
 			staticAttrs: lib.staticAttrs,
-			...defAttrs,
+			...(createElement || lib.js) ? defAttrs : defInlineAttrs,
 			...lib.attrs
 		};
 
@@ -326,7 +325,7 @@ function getLinkDecl(link) {
 	const attrs = normalizeAttrs({
 		href: link.src,
 		staticAttrs: link.staticAttrs,
-		...defAttrs,
+		...link.js ? defAttrs : defInlineAttrs,
 		...link.attrs
 	}, link.js);